src/requirements.txt
fastapi==0.66.0
uvicorn==0.14.0
asyncpg==0.23.0
tortoise-orm==0.17.5